# PERF: Micro optimize find_common_type