The `find_common_type` is a (small) performance bottleneck in one of the scipy algoritms. This PR optimizes the method:

A benchmark (extracted from  what is called from scipy):
```
import numpy as np
from numpy.core.numerictypes import find_common_type

array_types = [np.dtype('float64'), np.dtype('float64')]
scalar_types = []

find_common_type(array_types, scalar_types)
```
Results of `%timeit find_common_type(array_types, scalar_types)`.

* Main: `8.37 µs ± 500 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)`
* PR: `1.09 µs ± 57.6 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)` 
* PR (only first commit): `1.81 µs ± 313 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)`  

The optmizations are:
* Cache cast of  `__test_types` to `dtype`
* Make elements of  `dtypelist` unique. This could make other calls a bit slower, but the `list(set(dtypelist))` is fast and it makes cases with duplicate dtypes faster
* Fast path in `find_common_type` for empty arguments.





<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->