@mattip 2022-06-13
Can we suggest a fix to scipy to avoid this function that is also more performant?