@mattip 2022-06-13
Ahh, you already did :). Since the alternatives are more performant, I think we should close this and deprecate `find_common_type`