@eendebakpt 2022-06-19
Closing this PR. The issue has been solved in scipy.